ticket-kafka/ticket-be/requirements.txt
redis==3.2.1
kafka-python==1.4.6
